test(JestMigration): convert RadialBarChart
