+++ test-jest/chart/RadialBarChart.spec.tsx
@@ -1,0 +1,229 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+
+import { RadialBarChart, RadialBar, Legend, Sector, Tooltip } from '../../src';
+
+describe('<RadialBarChart />', () => {
+  const data = [
+    { name: '18-24', uv: 31.47, pv: 2400, fill: '#8884d8' },
+    { name: '25-29', uv: 26.69, pv: 4567, fill: '#83a6ed' },
+    { name: '30-34', uv: 15.69, pv: 1398, fill: '#8dd1e1' },
+    { name: '35-39', uv: 8.22, pv: 9800, fill: '#82ca9d' },
+    { name: '40-49', uv: 8.63, pv: 3908, fill: '#a4de6c' },
+    { name: '50+', uv: 2.63, pv: 4800, fill: '#d0ed57' },
+    { name: 'unknown', uv: 6.67, pv: 4800, fill: '#ffc658' },
+  ];
+
+  it('Renders 7 sectors in simple RadialBarChart', () => {
+    const label = { orientation: 'outer' };
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={data}
+      >
+        <RadialBar label={label} background dataKey="uv" isAnimationActive={false} />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelectorAll('.recharts-radial-bar-sector')).toHaveLength(data.length);
+  });
+
+  it("Don't renders any sectors when no RadialBar is added", () => {
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={data}
+      />,
+    );
+
+    expect(container.querySelectorAll('.recharts-radial-bar-sector')).toHaveLength(0);
+  });
+
+  it("Don't renders any sectors when width or height is smaller than 0", () => {
+    const { container } = render(
+      <RadialBarChart
+        width={0}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={data}
+      >
+        <RadialBar background dataKey="uv" isAnimationActive={false} />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelectorAll('.recharts-radial-bar-sector')).toHaveLength(0);
+  });
+
+  it('Renders sectors when barSize is not specified', () => {
+    const label = { orientation: 'outer' };
+    const { container } = render(
+      <RadialBarChart width={500} height={300} cx={150} cy={150} innerRadius={20} outerRadius={140} data={data}>
+        <RadialBar label={label} background dataKey="uv" isAnimationActive={false} />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelectorAll('.recharts-radial-bar-sector')).toHaveLength(data.length);
+  });
+
+  it('Renders 7 sectors in simple RadialBarChart', () => {
+    const label = { orientation: 'outer' };
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={data}
+      >
+        <RadialBar startAngle={0} endAngle={180} label={label} background dataKey="uv" isAnimationActive={false} />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelectorAll('.recharts-radial-bar-sector')).toHaveLength(data.length);
+  });
+
+  it('Renders 7 customized shape when shape is set to be a react element', () => {
+    const label = { orientation: 'outer' };
+    const Shape = (props: any) => <Sector {...props} className="customized-shape" />;
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={data}
+      >
+        <RadialBar shape={<Shape />} label={label} background dataKey="uv" isAnimationActive={false} />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelectorAll('.customized-shape')).toHaveLength(7);
+  });
+
+  it('Renders 7 customized shape when shape is set to be a function', () => {
+    const label = { orientation: 'outer' };
+    const renderShape = (props: any) => <Sector {...props} className="customized-shape" />;
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={data}
+      >
+        <RadialBar shape={renderShape} label={label} background dataKey="uv" isAnimationActive={false} />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelectorAll('.customized-shape')).toHaveLength(7);
+  });
+
+  it('Renders 7 legend item when add a Legend element', () => {
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={data}
+      >
+        <RadialBar background dataKey="uv" isAnimationActive={false} />
+        <Legend />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelectorAll('.recharts-legend-wrapper')).toHaveLength(1);
+    expect(container.querySelectorAll('.recharts-legend-item')).toHaveLength(7);
+  });
+
+  it('Should fill the RadialBar with color when fill property is present in data', () => {
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={[data[0]]}
+      >
+        <RadialBar background dataKey="uv" isAnimationActive={false} />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelector('.recharts-radial-bar-sector')).toHaveAttribute('fill', data[0].fill);
+  });
+
+  it('Should fill the RadialBar with the default color when fill property is not present in data', () => {
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={[{ name: '18-24', uv: 31.47, pv: 2400 }]}
+      >
+        <RadialBar background dataKey="uv" isAnimationActive={false} />
+      </RadialBarChart>,
+    );
+
+    expect(container.querySelector('.recharts-radial-bar-sector')).not.toHaveAttribute('fill');
+  });
+
+  it('Renders tooltip when add a Tooltip element', async () => {
+    const { container } = render(
+      <RadialBarChart
+        width={500}
+        height={300}
+        cx={150}
+        cy={150}
+        innerRadius={20}
+        outerRadius={140}
+        barSize={10}
+        data={data}
+      >
+        <RadialBar background dataKey="uv" isAnimationActive={false} />
+        <Tooltip />
+      </RadialBarChart>,
+    );
+
+    const user = userEvent.setup();
+    user.hover(container);
+
+    expect(container.querySelector('.recharts-default-tooltip')).toBeInTheDocument();
+    expect(container.querySelector('.recharts-tooltip-wrapper')).toBeInTheDocument();
+  });
+});
